Increase inode count for system.img
Bliss now ships with more packages, which hits against the inode
limit for bullhead. The inode limit was just set to the bare
minimum suitable for AOSP. That means we only can use about 1.5 GB
out of the total 3 GB (3072 MB) allocated for the system image.
While this does significantly cut down on image size, unused space
is wasted space, and we really want to fill that other 1.5 GB with
usable content.